LockedIiifResource / getLockedIiifResource
